refactoring to remove internal tracking error
